+++ src/GafferSceneUI/ParameterInspector.cpp
@@ -157,8 +157,12 @@ Gaffer::ValuePlugPtr ParameterInspector::source( const GafferScene::SceneAlgo::H
 
 		for( const auto &tweak : TweakPlug::Range( *shaderTweaks->tweaksPlug() ) )
 		{
-			/// \todo Consider shader as well as name
-			if( tweak->namePlug()->getValue() == m_parameter.name.string() && tweak->enabledPlug()->getValue() )
+			const std::string tweakName = (
+				m_parameter.shader.string() +
+				( m_parameter.shader.string().empty() ? "" : "." ) +
+				m_parameter.name.string()
+			);
+			if( tweak->namePlug()->getValue() == tweakName && tweak->enabledPlug()->getValue() )
 			{
 				return tweak;
 			}