+++ Changes.md
@@ -9,7 +9,7 @@ Features
 Improvements
 ------------
 
-- LightEditor : Added the ability to register columns for editing any shader in the light's shader network with an Edit Scope. The parameter to edit is registered in the form `shaderName.parameterName` when calling `GafferSceneUI.LightEditor.registerParameter`.
+- LightEditor : Added the ability to register columns for editing any shader in the light's shader network with an Edit Scope or a `ShaderTweaks` node. The parameter to edit is registered in the form `shaderName.parameterName` when calling `GafferSceneUI.LightEditor.registerParameter`.
 
 1.4.2.0 (relative to 1.4.1.0)
 =======